# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminDocumentView": {
        "properties": {
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "document_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Number"
          },
          "document_type": {
            "title": "Document Type",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "holder_name": {
            "title": "Holder Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "internal_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Notes"
          },
          "issued_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "provider_display_name": {
            "title": "Provider Display Name",
            "type": "string"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/DocumentReviewView"
            },
            "title": "Reviews",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "submitted_at": {
            "format": "date-time",
            "title": "Submitted At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_type",
          "document_number",
          "holder_name",
          "issuer",
          "jurisdiction",
          "issued_at",
          "expires_at",
          "status",
          "submitted_at",
          "reviewed_at",
          "rejection_reason",
          "filename",
          "mime_type",
          "byte_size",
          "reviews",
          "provider_id",
          "provider_display_name",
          "reviewed_by",
          "internal_notes"
        ],
        "title": "AdminDocumentView",
        "type": "object"
      },
...
      "DocumentReviewView": {
        "properties": {
          "actor_kind": {
            "$ref": "#/components/schemas/ReviewActorKind"
          },
          "audit_reference": {
            "format": "uuid",
            "title": "Audit Reference",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "new_status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "previous_status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "id",
          "actor_kind",
          "previous_status",
          "new_status",
          "reason",
          "audit_reference",
          "created_at"
        ],
        "title": "DocumentReviewView",
        "type": "object"
      },
      "DocumentStatus": {
        "enum": [
          "DRAFT",
          "UPLOADED",
          "SCANNING",
          "PENDING",
          "IN_REVIEW",
          "OBSERVED",
          "APPROVED",
          "REJECTED",
          "EXPIRED",
          "SUSPENDED"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PricingType": {
        "enum": [
          "FIXED",
          "FROM",
          "QUOTE",
          "HOURLY",
          "PER_SESSION",
          "PER_UNIT"
        ],
        "title": "PricingType",
        "type": "string"
      },
...
      "PublicReviewView": {
        "properties": {
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "service_name": {
            "title": "Service Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "service_name",
          "rating",
          "comment",
          "published_at"
        ],
        "title": "PublicReviewView",
        "type": "object"
      },
...
      "ReviewCreate": {
        "properties": {
          "comment": {
            "maxLength": 2000,
            "minLength": 10,
            "title": "Comment",
            "type": "string"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          }
        },
        "required": [
          "rating",
          "comment"
        ],
        "title": "ReviewCreate",
        "type": "object"
      },
      "ReviewDecision": {
        "properties": {
          "action": {
            "enum": [
              "publish",
              "reject",
              "hide"
            ],
            "title": "Action",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "action"
        ],
        "title": "ReviewDecision",
        "type": "object"
      },
      "ReviewStatus": {
        "enum": [
          "PENDING",
          "PUBLISHED",
          "REJECTED",
          "HIDDEN"
        ],
        "title": "ReviewStatus",
        "type": "string"
      },
      "ReviewView": {
        "properties": {
          "booking_id": {
            "format": "uuid",
            "title": "Booking Id",
            "type": "string"
          },
          "client_id": {
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "moderation_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation Reason"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "provider_name": {
            "title": "Provider Name",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "rating": {
            "title": "Rating",
            "type": "integer"
          },
          "service_name": {
            "title": "Service Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "booking_id",
          "client_id",
          "provider_id",
          "service_name",
          "client_name",
          "provider_name",
          "rating",
          "comment",
          "status",
          "moderation_reason",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewView",
        "type": "object"
      },
...
      "ServiceRequestView": {
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/RequestAttachmentView"
            },
            "title": "Attachments",
            "type": "array"
          },
          "booking_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Id"
          },
          "client_address_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Address Label"
          },
          "client_id": {
            "format": "uuid",
            "title": "Client Id",
            "type": "string"
          },
          "client_name": {
            "title": "Client Name",
            "type": "string"
          },
          "configured_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configured Price"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "offer_headline": {
            "title": "Offer Headline",
            "type": "string"
          },
          "preferred_start_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Start At"
          },
          "price_currency": {
            "title": "Price Currency",
            "type": "string"
          },
          "pricing_type": {
            "$ref": "#/components/schemas/PricingType"
          },
          "provider_id": {
            "format": "uuid",
            "title": "Provider Id",
            "type": "string"
          },
          "provider_name": {
            "title": "Provider Name",
            "type": "string"
          },
          "provider_service_id": {
            "format": "uuid",
            "title": "Provider Service Id",
            "type": "string"
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteView"
            },
            "title": "Quotes",
            "type": "array"
          },
          "selected_modality": {
            "$ref": "#/components/schemas/ProviderModality"
          },
          "service_name": {
            "title": "Service Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ServiceRequestStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "viewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewed At"
          }
        },
        "required": [
          "id",
          "client_id",
          "provider_id",
          "provider_service_id",
          "service_name",
          "offer_headline",
          "pricing_type",
          "configured_price",
          "price_currency",
          "client_name",
          "provider_name",
          "selected_modality",
          "client_address_label",
          "title",
          "description",
          "preferred_start_at",
          "status",
          "viewed_at",
          "created_at",
          "attachments",
          "quotes",
          "booking_id"
        ],
        "title": "ServiceRequestView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SIC API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/documents/{document_id}/review": {
      "post": {
        "operationId": "begin_document_review_v1_admin_documents__document_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDocumentView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Begin Document Review",
        "tags": [
          "admin-documents"
        ]
      }
    },
...
    "/v1/admin/reviews": {
      "get": {
        "operationId": "admin_reviews_v1_admin_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ReviewStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Statuses"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewView"
                  },
                  "title": "Response Admin Reviews V1 Admin Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Reviews",
        "tags": [
          "admin-reviews"
        ]
      }
    },
    "/v1/admin/reviews/{review_id}/moderate": {
      "post": {
        "operationId": "moderate_review_v1_admin_reviews__review_id__moderate_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Moderate Review",
        "tags": [
          "admin-reviews"
        ]
      }
    },
...
    "/v1/client/reviews": {
      "get": {
        "operationId": "client_reviews_v1_client_reviews_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewView"
                  },
                  "title": "Response Client Reviews V1 Client Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Client Reviews",
        "tags": [
          "client-reviews"
        ]
      }
    },
    "/v1/client/reviews/bookings/{booking_id}": {
      "post": {
        "operationId": "submit_review_v1_client_reviews_bookings__booking_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "booking_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Booking Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Review",
        "tags": [
          "client-reviews"
        ]
      }
    },
    "/v1/client/reviews/{review_id}": {
      "patch": {
        "operationId": "update_review_v1_client_reviews__review_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Review",
        "tags": [
          "client-reviews"
        ]
      }
    },
...
    "/v1/provider/reviews": {
      "get": {
        "operationId": "provider_reviews_v1_provider_reviews_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewView"
                  },
                  "title": "Response Provider Reviews V1 Provider Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Provider Reviews",
        "tags": [
          "provider-reviews"
        ]
      }
    },
...
    "/v1/provider/service-requests/{request_id}/view": {
      "post": {
        "operationId": "view_request_v1_provider_service_requests__request_id__view_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRequestView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "View Request",
        "tags": [
          "provider-engagements"
        ]
      }
    },
...
    "/v1/providers/{provider_slug}/reviews": {
      "get": {
        "operationId": "public_reviews_v1_providers__provider_slug__reviews_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider_slug",
            "required": true,
            "schema": {
              "title": "Provider Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicReviewView"
                  },
                  "title": "Response Public Reviews V1 Providers  Provider Slug  Reviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Reviews",
        "tags": [
          "public-reviews"
        ]
      }
    },
    "/v1/providers/{slug}": {
      "get": {
        "operationId": "public_provider_v1_providers__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProviderProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Provider",
        "tags": [
          "public-search"
        ]
      }
    },
...
  }
}